refactor(ui): rewire cell_editor and document_preview to composites::ModalFrame (#102)
Move CellEditorModal and DocumentPreviewModal into dbflux_components::modals.
Rewire both from the dbflux_ui shim to dbflux_components::composites::ModalFrame
directly, reproducing the three shim defaults: key_context(SqlPreviewModal),
close_icon(X), header_leading(Pencil/Braces). Replace source files with shims.

Remove overlays/cell_editor_modal.rs and overlays/document_preview_modal.rs
from the guardrail test path list: they no longer flow through the shim by
design (they use the composite directly). The five remaining entries are
unchanged.
