+++ crates/dbflux_components/src/modals/mod.rs
@@ -1,5 +1,7 @@
 pub mod active_query;
+pub mod cell_editor;
 pub mod delete_connection;
+pub mod document_preview;
 pub mod drop_table;
 pub mod schema_drift;
 pub mod shell;
@@ -7,7 +9,9 @@ pub mod tunnel_auth;
 pub mod unsaved_changes;
 
 pub use active_query::{ActiveQueryOutcome, ActiveQueryRequest, ActiveQueryTrigger, ModalActiveQuery};
+pub use cell_editor::{CellEditorClosedEvent, CellEditorModal, CellEditorSaveEvent};
 pub use delete_connection::{DeleteConnectionOutcome, DeleteConnectionRequest, ModalDeleteConnection};
+pub use document_preview::{DOC_INDEX_NEW, DocumentPreviewClosedEvent, DocumentPreviewModal, DocumentPreviewSaveEvent};
 pub use drop_table::{DropTableOutcome, DropTableRequest, ModalDropTable};
 pub use schema_drift::{ModalSchemaDrift, SchemaDriftContinue, SchemaDriftDismissed, SchemaDriftRefresh};
 pub use shell::{ModalShell, ModalVariant};
